printf corregido

+++ printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: alfrberm <alfrberm@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/07 15:11:15 by alfrberm          #+#    #+#             */
-/*   Updated: 2024/10/07 16:34:54 by alfrberm         ###   ########.fr       */
+/*   Updated: 2024/10/08 08:22:40 by alfrberm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	ft_putchar(int c);
 int	ft_putstr(const char *str);
 int	ft_putnbr(int nb);
 int	ft_putnbr_unsigned(unsigned int nb);
-int	ft_putnbr_base(int nb, const char *base);
+int	ft_putnbr_base(unsigned long nb, const char *base);
 int	ft_putptr(void *ptr);
 
 #endif
